chore: remove components from use activity details hook
