+++ packages/client/components/ActivityLibrary/ActivityDetails/hooks/useActivityDetails.tsx
@@ -1,17 +1,9 @@
-import React, {useState} from 'react'
+import React from 'react'
 import {ActivityDetailsQuery} from '~/__generated__/ActivityDetailsQuery.graphql'
 import {CategoryID, QUICK_START_CATEGORY_ID} from '../../Categories'
 import {MeetingTypeEnum} from '../../../../../server/postgres/types/Meeting'
 import {ActivityId, getActivityIllustration} from '../../ActivityIllustrations'
 import {useHistory} from 'react-router'
-import {TemplateDetails} from '../TemplateDetails'
-import EditableTemplateName from '../../../../modules/meeting/components/EditableTemplateName'
-import {MeetingDetails} from '../MeetingDetails'
-import ActivityDetailsSidebar from '../../ActivityDetailsSidebar'
-import AddPokerTemplateDimension from '../../../../modules/meeting/components/AddPokerTemplateDimension'
-import AddTemplatePrompt from '../../../../modules/meeting/components/AddTemplatePrompt'
-import TemplateDimensionList from '../../../../modules/meeting/components/TemplateDimensionList'
-import TemplatePromptList from '../../../../modules/meeting/components/TemplatePromptList'
 
 export type NoTemplatesMeeting = Exclude<MeetingTypeEnum, 'retrospective' | 'poker'>
 
@@ -62,28 +54,32 @@ export const MEETING_TYPE_TIP_LOOKUP: Record<MeetingTypeEnum, React.ReactNode> =
   poker: <>sync estimations with your backlog</>
 }
 
-export interface Activity {
+type MeetingActivity = {
+  isTemplate: false
+}
+
+type TemplateActivity = {
+  isTemplate: true
+  template: ActivityDetailsQuery['response']['viewer']['availableTemplates']['edges'][number]['node']
+}
+
+export type Activity = {
   id: ActivityId
-  name: React.ReactNode
+  name: string
   illustration: string
   usageStats?: any
   category: CategoryID
   categoryLink: string
   type: MeetingTypeEnum
   description: React.ReactNode
-  details: React.ReactNode
-  isTemplate: boolean
-  templateConfig?: React.ReactNode
-  sidebar?: React.ReactNode
-}
+} & (MeetingActivity | TemplateActivity)
 
 export const useActivityDetails = (
   activityIdParam: string,
   data: ActivityDetailsQuery['response']
 ) => {
   const {viewer} = data
-  const {availableTemplates, teams} = viewer
-  const [isEditing, setIsEditing] = useState(false)
+  const {availableTemplates} = viewer
 
   const history = useHistory<{prevCategory?: string; edit?: boolean}>()
   const prevCategory = history.location.state?.prevCategory
@@ -96,11 +92,7 @@ export const useActivityDetails = (
     const category = MEETING_TYPE_TO_CATEGORY_ID[type]
     const activity: Activity = {
       id: activityId,
-      name: (
-        <div className='text-[32px] font-semibold leading-tight text-slate-700'>
-          {MEETING_ID_TO_NAME[type]}
-        </div>
-      ),
+      name: MEETING_ID_TO_NAME[type],
       illustration: getActivityIllustration(activityId),
       category,
       categoryLink: `/activity-library/category/${
@@ -108,13 +100,10 @@ export const useActivityDetails = (
       }`,
       type: activityId as MeetingTypeEnum,
       description: MEETING_TYPE_DESCRIPTION_LOOKUP[type],
-      isTemplate: false,
-      details: <MeetingDetails type={type} category={category} />
+      isTemplate: false
     }
 
     return {
-      isEditing,
-      setIsEditing,
       activity
     }
   }
@@ -122,61 +111,15 @@ export const useActivityDetails = (
   const selectedTemplate = availableTemplates.edges.find(
     (edge) => edge.node.id === activityIdParam
   )?.node
-  if (!selectedTemplate) return {isEditing, setIsEditing, activity: null}
+  if (!selectedTemplate) return {activity: null}
 
   const activityId = selectedTemplate.id as ActivityId
   const category = selectedTemplate.category as CategoryID
   const type = selectedTemplate.type
-  const teamTemplates = availableTemplates.edges
-    .map((edge) => edge.node)
-    .filter((edge) => edge.teamId === selectedTemplate.teamId)
-  const teamIds = teams.map((team) => team.id)
-  const isOwner = teamIds.includes(selectedTemplate.teamId)
-
-  const templateConfigLookup: Record<MeetingTypeEnum, React.ReactNode> = {
-    retrospective: (
-      <>
-        <TemplatePromptList
-          isOwner={isOwner && isEditing}
-          prompts={selectedTemplate.prompts!}
-          templateId={selectedTemplate.id}
-        />
-        {isOwner && isEditing && (
-          <AddTemplatePrompt templateId={selectedTemplate.id} prompts={selectedTemplate.prompts!} />
-        )}
-      </>
-    ),
-    poker: (
-      <>
-        <TemplateDimensionList
-          isOwner={isOwner}
-          readOnly={!isEditing}
-          dimensions={selectedTemplate.dimensions!}
-          templateId={selectedTemplate.id}
-        />
-        {isOwner && isEditing && (
-          <AddPokerTemplateDimension
-            templateId={selectedTemplate.id}
-            dimensions={selectedTemplate.dimensions!}
-          />
-        )}
-      </>
-    ),
-    action: null,
-    teamPrompt: null
-  }
 
   const activity: Activity = {
     id: activityId,
-    name: (
-      <EditableTemplateName
-        className='text-[32px] leading-9'
-        name={selectedTemplate.name}
-        templateId={selectedTemplate.id}
-        teamTemplates={teamTemplates}
-        isOwner={isOwner && isEditing}
-      />
-    ),
+    name: selectedTemplate.name,
     illustration: getActivityIllustration(activityId),
     category,
     categoryLink: `/activity-library/category/${
@@ -185,28 +128,10 @@ export const useActivityDetails = (
     type,
     description: MEETING_TYPE_DESCRIPTION_LOOKUP[type],
     isTemplate: true,
-    details: (
-      <TemplateDetails
-        selectedTemplate={selectedTemplate}
-        viewerRef={viewer}
-        templatesRef={availableTemplates.edges.map((edge) => edge.node)}
-        isEditing={isEditing}
-        setIsEditing={setIsEditing}
-      />
-    ),
-    templateConfig: templateConfigLookup[type],
-    sidebar: (
-      <ActivityDetailsSidebar
-        selectedTemplateRef={selectedTemplate}
-        teamsRef={teams}
-        isOpen={!isEditing}
-      />
-    )
+    template: selectedTemplate
   }
 
   return {
-    isEditing,
-    setIsEditing,
     activity
   }
 }
